Add auto-allocation for bgpvpn creation endpoint.
This commit adds auto-allocation system for bgpvpn creation. If the
special setting are enabled the service plugin will allocate
import/export/route targets for new bgpvpn in case if they were not
passed in request body.

The algorithm how targets will be choosen:
1. 4-byte ASn will be converted from dotted notation to full number
    by formula (<AS> * 65536) + <Number>. For example if setting
   `region_asn` has value 65130.4 the result ASn will be (65130 *
   65536) + 4 = 4268359684;
2. unique identifier for this bgpvpn will be found from the range of
    option `target_id_range`;
3. the ASn from first step will be used for import/export/route fields
    depending on what option was enabled from the list:
    import_target_auto_allocation, export_target_auto_allocation,
    route_target_auto_allocation and unique identifier will be added
    after colon like: 4268359684:100.

As result, we will have unique auto-allocated targets when the option
enabled.

Change-Id: I46cf01d0f4005b885e799964c7bd6c27efcc5c2d